['Activities'] redesigned it: offer 2 nav tabs 'All Activities' & 'My Activities'; offer 'Modified by' user filter
