+++ frontend/src/pages/dashboard/files-activities.js
@@ -1,306 +1,25 @@
 import React, { Component, Fragment } from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
+import { Link } from '@gatsbyjs/reach-router';
 import { seafileAPI } from '../../utils/seafile-api';
-import { gettext, siteRoot, serviceURL } from '../../utils/constants';
+import { gettext, siteRoot, username } from '../../utils/constants';
 import { Utils } from '../../utils/utils';
 import Loading from '../../components/loading';
 import Activity from '../../models/activity';
-import ListCreatedFileDialog from '../../components/dialog/list-created-files-dialog';
-import ModalPortal from '../../components/modal-portal';
+import FileActivitiesContent from './content';
+import UserSelector from './user-selector';
 
 import '../../css/files-activities.css';
 
 moment.locale(window.app.config.lang);
 
-const contentPropTypes = {
-  isLoadingMore: PropTypes.bool.isRequired,
-  items: PropTypes.array.isRequired,
+const propTypes = {
+  onlyMine: PropTypes.bool
 };
-
-class FileActivitiesContent extends Component {
-
-  render() {
-    const isDesktop = Utils.isDesktop();
-    let { items, isLoadingMore } = this.props;
-
-    const desktopThead = (
-      <thead>
-        <tr>
-          <th width="8%">{/* avatar */}</th>
-          <th width="15%">{gettext('User')}</th>
-          <th width="20%">{gettext('Operation')}</th>
-          <th width="37%">{gettext('File')} / {gettext('Library')}</th>
-          <th width="20%">{gettext('Time')}</th>
-        </tr>
-      </thead>
-    );
-
-    const mobileThead = (
-      <thead>
-        <tr>
-          <th width="15%"></th>
-          <th width="53%"></th>
-          <th width="32%"></th>
-        </tr>
-      </thead>
-    );
-
-    return (
-      <Fragment>
-        <table className="table-hover table-thead-hidden">
-          {isDesktop ? desktopThead : mobileThead}
-          <tbody>
-            {items.map((item, index) => {
-              return (
-                <ActivityItem
-                  key={index}
-                  isDesktop={isDesktop}
-                  item={item}
-                  index={index}
-                  items={items}
-                />
-              );
-            })}
-          </tbody>
-        </table>
-        {isLoadingMore ? <span className="loading-icon loading-tip"></span> : ''}
-      </Fragment>
-    );
-  }
-}
-
-FileActivitiesContent.propTypes = contentPropTypes;
-
-const activityPropTypes = {
-  item: PropTypes.object.isRequired,
-  index: PropTypes.number.isRequired,
-  items: PropTypes.array.isRequired,
-  isDesktop: PropTypes.bool.isRequired,
-};
-
-class ActivityItem extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      isListCreatedFiles: false,
-    };
-  }
-
-  onListCreatedFilesToggle = () => {
-    this.setState({
-      isListCreatedFiles: !this.state.isListCreatedFiles,
-    });
-  };
-
-  render() {
-    const isDesktop = this.props.isDesktop;
-    let {item, index, items} = this.props;
-    let op, details, moreDetails = false;
-    let userProfileURL = `${siteRoot}profile/${encodeURIComponent(item.author_email)}/`;
-
-    let libURL = siteRoot + 'library/' + item.repo_id + '/' + encodeURIComponent(item.repo_name) + '/';
-    let libLink = <a href={libURL}>{item.repo_name}</a>;
-    let smallLibLink = <a className="small text-secondary" href={libURL}>{item.repo_name}</a>;
-
-    if (item.obj_type == 'repo') {
-      switch(item.op_type) {
-        case 'create':
-          op = gettext('Created library');
-          details = libLink;
-          break;
-        case 'rename':
-          op = gettext('Renamed library');
-          details = <span>{item.old_repo_name} => {libLink}</span>;
-          break;
-        case 'delete':
-          op = gettext('Deleted library');
-          details = item.repo_name;
-          break;
-        case 'recover':
-          op = gettext('Restored library');
-          details = libLink;
-          break;
-        case 'clean-up-trash':
-          op = gettext('Cleaned trash');
-          if (item.days == 0) {
-            details = gettext('Removed all items from trash.');
-          } else {
-            details = gettext('Removed items older than {n} days from trash.').replace('{n}', item.days);
-          }
-          moreDetails = true;
-          break;
-      }
-    } else if (item.obj_type == 'draft') {
-      let fileURL = `${siteRoot}lib/${item.repo_id}/file${Utils.encodePath(item.path)}`;
-      let fileLink = <a href={fileURL} target="_blank" rel="noreferrer">{item.name}</a>;
-      op = gettext('Publish draft');
-      details = fileLink;
-      moreDetails = true;
-    } else if (item.obj_type == 'files') {
-      let fileURL = `${siteRoot}lib/${item.repo_id}/file${Utils.encodePath(item.path)}`;
-      if (item.name.endsWith('(draft).md')) {
-        fileURL = serviceURL + '/drafts/' + item.draft_id + '/';
-      }
-      let fileLink = `<a href=${fileURL} target="_blank">${item.name}</a>`;
-      if (item.name.endsWith('(draft).md') && !item.draft_id) {
-        fileLink = item.name;
-      }
-      let fileCount = item.createdFilesCount - 1;
-      let firstLine = gettext('{file} and {n} other files')
-        .replace('{file}', fileLink)
-        .replace('{n}', fileCount);
-      op = gettext('Created {n} files').replace('{n}', item.createdFilesCount);
-      details = (
-        <Fragment>
-          <p className="m-0 d-inline" dangerouslySetInnerHTML={{__html: firstLine}}></p>
-          {isDesktop && <button type="button" onClick={this.onListCreatedFilesToggle} className="activity-details text-secondary ml-2 border-0 p-0 bg-transparent">{gettext('details')}</button>}
-        </Fragment>
-      );
-      moreDetails = true;
-    } else if (item.obj_type == 'file') {
-      const isDraft = item.name.endsWith('(draft).md');
-      const fileURL = isDraft ? serviceURL + '/drafts/' + item.draft_id + '/' :
-        `${siteRoot}lib/${item.repo_id}/file${Utils.encodePath(item.path)}`;
-      let fileLink = <a href={fileURL} target="_blank" rel="noreferrer">{item.name}</a>;
-      if (isDraft && !item.draft_id) {
-        fileLink = item.name;
-      }
-      switch (item.op_type) {
-        case 'create':
-          op = isDraft ? gettext('Created draft') : gettext('Created file');
-          details = fileLink;
-          moreDetails = true;
-          break;
-        case 'delete':
-          op = isDraft ? gettext('Deleted draft') : gettext('Deleted file');
-          details = item.name;
-          moreDetails = true;
-          break;
-        case 'recover':
-          op = gettext('Restored file');
-          details = fileLink;
-          moreDetails = true;
-          break;
-        case 'rename':
-          op = gettext('Renamed file');
-          details = <span>{item.old_name} => {fileLink}</span>;
-          moreDetails = true;
-          break;
-        case 'move':
-          // eslint-disable-next-line
-          const filePathLink = <a href={fileURL}>{item.path}</a>;
-          op = gettext('Moved file');
-          details = <span>{item.old_path} => {filePathLink}</span>;
-          moreDetails = true;
-          break;
-        case 'edit': // update
-          op = isDraft ? gettext('Updated draft') : gettext('Updated file');
-          details = fileLink;
-          moreDetails = true;
-          break;
-      }
-    } else { // dir
-      let dirURL = siteRoot + 'library/' + item.repo_id + '/' + encodeURIComponent(item.repo_name) + Utils.encodePath(item.path);
-      let dirLink = <a href={dirURL} target="_blank" rel="noreferrer">{item.name}</a>;
-      switch (item.op_type) {
-        case 'create':
-          op = gettext('Created folder');
-          details = dirLink;
-          moreDetails = true;
-          break;
-        case 'delete':
-          op = gettext('Deleted folder');
-          details = item.name;
-          moreDetails = true;
-          break;
-        case 'recover':
-          op = gettext('Restored folder');
-          details = dirLink;
-          moreDetails = true;
-          break;
-        case 'rename':
-          op = gettext('Renamed folder');
-          details = <span>{item.old_name} => {dirLink}</span>;
-          moreDetails = true;
-          break;
-        case 'move':
-          // eslint-disable-next-line
-          const dirPathLink = <a href={dirURL}>{item.path}</a>;
-          op = gettext('Moved folder');
-          details = <span>{item.old_path} => {dirPathLink}</span>;
-          moreDetails = true;
-          break;
-      }
-    }
-
-    let isShowDate = true;
-    if (index > 0) {
-      let lastEventTime = items[index - 1].time;
-      isShowDate = moment(item.time).isSame(lastEventTime, 'day') ? false : true;
-    }
-
-    return (
-      <Fragment>
-        {isShowDate &&
-          <tr>
-            <td colSpan={isDesktop ? 5 : 3} className="border-top-0">{moment(item.time).format('YYYY-MM-DD')}</td>
-          </tr>
-        }
-        {isDesktop ? (
-          <tr>
-            <td className="text-center">
-              <img src={item.avatar_url} alt="" width="32" height="32" className="avatar" />
-            </td>
-            <td>
-              <a href={userProfileURL}>{item.author_name}</a>
-            </td>
-            <td>{op}</td>
-            <td>
-              {details}
-              {moreDetails && <br /> }
-              {moreDetails && smallLibLink}
-            </td>
-            <td className="text-secondary">
-              <time datetime={item.time} is="relative-time" title={moment(item.time).format('llll')}>{moment(item.time).fromNow()}</time>
-            </td>
-          </tr>
-        ) : (
-          <tr>
-            <td className="text-center align-top">
-              <img src={item.avatar_url} alt="" width="32" height="32" className="avatar" />
-            </td>
-            <td>
-              <a href={userProfileURL}>{item.author_name}</a>
-              <p className="m-0 text-secondary">{op}</p>
-              {details}
-            </td>
-            <td className="text-right align-top">
-              <span className="text-secondary mobile-activity-time">
-                <time datetime={item.time} is="relative-time" title={moment(item.time).format('llll')}>{moment(item.time).fromNow()}</time>
-              </span>
-              {moreDetails && <br /> }
-              {moreDetails && libLink}
-            </td>
-          </tr>
-        )}
-        {this.state.isListCreatedFiles &&
-          <ModalPortal>
-            <ListCreatedFileDialog
-              activity={item}
-              toggleCancel={this.onListCreatedFilesToggle}
-            />
-          </ModalPortal>
-        }
-      </Fragment>
-    );
-  }
-}
-
-ActivityItem.propTypes = activityPropTypes;
 
 class FilesActivities extends Component {
+
   constructor(props) {
     super(props);
     this.state = {
@@ -309,21 +28,37 @@ class FilesActivities extends Component {
       isLoadingMore: false,
       currentPage: 1,
       hasMore: true,
+      allItems: [],
       items: [],
+      availableUsers: [],
+      targetUsers: []
     };
     this.avatarSize = 72;
     this.curPathList = [];
     this.oldPathList = [];
+    this.availableUserEmails = new Set();
   }
 
   componentDidMount() {
-    let currentPage = this.state.currentPage;
+    let { currentPage, availableUsers } = this.state;
     seafileAPI.listActivities(currentPage, this.avatarSize).then(res => {
       // {"events":[...]}
       let events = this.mergePublishEvents(res.data.events);
       events = this.mergeFileCreateEvents(events);
+      events.forEach(item => {
+        if (!this.availableUserEmails.has(item.author_email)) {
+          this.availableUserEmails.add(item.author_email);
+          availableUsers.push({
+            email: item.author_email,
+            name: item.author_name,
+            avatar_url: item.avatar_url
+          });
+        }
+      });
       this.setState({
-        items: events,
+        allItems: events,
+        items: this.filterEvents(events),
+        availableUsers: availableUsers,
         currentPage: currentPage + 1,
         isFirstLoading: false,
         hasMore: true,
@@ -400,19 +135,34 @@ class FilesActivities extends Component {
   };
 
   getMore() {
-    let currentPage = this.state.currentPage;
+    const { currentPage, availableUsers, targetUsers } = this.state;
     seafileAPI.listActivities(currentPage, this.avatarSize).then(res => {
       // {"events":[...]}
       let events = this.mergePublishEvents(res.data.events);
       events = this.mergeFileCreateEvents(events);
+      events.forEach(item => {
+        if (!this.availableUserEmails.has(item.author_email)) {
+          this.availableUserEmails.add(item.author_email);
+          availableUsers.push({
+            email: item.author_email,
+            name: item.author_name,
+            avatar_url: item.avatar_url
+          });
+        }
+      });
+      const filteredEvents = this.filterEvents(events);
       this.setState({
-        isLoadingMore: false,
-        items: [...this.state.items, ...events],
+        allItems: [...this.state.allItems, ...events],
+        items: [...this.state.items, ...filteredEvents],
+        availableUsers: availableUsers,
         currentPage: currentPage + 1,
+        isLoadingMore: false,
         hasMore: res.data.events.length === 0 ? false : true
       });
       if (this.state.items.length < 25 && this.state.hasMore) {
-        this.getMore();
+        if (!(targetUsers.length && currentPage == 100)) {
+          this.getMore();
+        }
       }
     }).catch(error => {
       this.setState({
@@ -421,6 +171,34 @@ class FilesActivities extends Component {
       });
     });
   }
+
+  filterEvents = (events) => {
+    const { onlyMine } = this.props;
+    const { targetUsers } = this.state;
+
+    if (onlyMine) {
+      return events.filter(item => item.author_email == username);
+    } else if (targetUsers.length) {
+      return events.filter(item => targetUsers.map(item => item.email).indexOf(item.author_email) != -1);
+    } else {
+      return events;
+    }
+  };
+
+  setTargetUsers = (selectedUsers) => {
+    this.setState({
+      targetUsers: selectedUsers
+    }, () => {
+      const items = this.filterEvents(this.state.allItems);
+      this.setState({
+        items: items
+      }, () => {
+        if (items.length < 25 && this.state.hasMore) {
+          this.getMore();
+        }
+      });
+    });
+  };
 
   handleScroll = (event) => {
     if (!this.state.isLoadingMore && this.state.hasMore) {
@@ -437,19 +215,38 @@ class FilesActivities extends Component {
   };
 
   render() {
+    const { onlyMine } = this.props;
+    const { targetUsers, availableUsers } = this.state;
     return (
       <div className="main-panel-center">
         <div className="cur-view-container" id="activities">
           <div className="cur-view-path">
-            <h3 className="sf-heading">{gettext('Activities')}</h3>
+            <ul className="nav">
+              <li className="nav-item">
+                <Link to={`${siteRoot}dashboard/`} className={`nav-link${onlyMine ? '' : ' active'}`}>{gettext('All Activities')}</Link>
+              </li>
+              <li className="nav-item">
+                <Link to={`${siteRoot}my-activities/`} className={`nav-link${onlyMine ? ' active': ''}`}>{gettext('My Activities')}</Link>
+              </li>
+            </ul>
           </div>
           <div className="cur-view-content d-block" onScroll={this.handleScroll}>
             {this.state.isFirstLoading && <Loading />}
             {(!this.state.isFirstLoading && this.state.errorMsg) &&
               <p className="error text-center">{this.state.errorMsg}</p>
             }
-            {!this.state.isFirstLoading &&
-              <FileActivitiesContent items={this.state.items} isLoadingMore={this.state.isLoadingMore}/>
+            {!this.state.isFirstLoading && (
+              <Fragment>
+                {!onlyMine && (
+                  <UserSelector
+                    availableUsers={availableUsers}
+                    currentSelectedUsers={targetUsers}
+                    setTargetUsers={this.setTargetUsers}
+                  />
+                )}
+                <FileActivitiesContent items={this.state.items} isLoadingMore={this.state.isLoadingMore} />
+              </Fragment>
+            )
             }
           </div>
         </div>
@@ -457,5 +254,7 @@ class FilesActivities extends Component {
     );
   }
 }
+
+FilesActivities.propTypes = propTypes;
 
 export default FilesActivities;
