+++ frontend/src/pages/dashboard/my-file-activities.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+import FilesActivities from './files-activities';
+
+class MyFilesActivities extends Component {
+
+  render() {
+    return <FilesActivities onlyMine={true} />;
+  }
+}
+
+export default MyFilesActivities;
